UserProfile in separate Component
